Harden and refactor computed object queries
