+++ src/db/traits/computed_field.rs
@@ -18,7 +18,7 @@ use crate::db::{DbConnection, DbPool, with_connection, with_transaction};
 use crate::errors::ApiError;
 use crate::events::{Action, EntityType, EventContext, NewEvent, emit_event};
 use crate::models::search::{
-    ComputedFieldScope, ComputedSortValueType, FilterField, Operator, ParsedQueryParam,
+    ComputedFieldScope, ComputedQueryValueType, FilterField, Operator, ParsedQueryParam,
     ParsedQueryParamExt, QueryOptions, SQLComponent, SQLValue, SortParam,
 };
 use crate::models::{
@@ -29,7 +29,28 @@ use crate::models::{
     HubuumObjectComputedResponse, NewObjectComputedData, NewTaskEventRecord, ObjectComputedData,
     SharedComputedScopeResponse, TaskKind, TaskRecord, TaskStatus, ValidatedComputedFieldPatch,
 };
-use crate::pagination::{CursorSqlMapping, CursorSqlType, OwnedCursorSqlField};
+use crate::pagination::{CursorSqlField, CursorSqlMapping, CursorSqlType};
+
+mod materialization;
+mod query;
+mod rebuild;
+
+pub(crate) use materialization::materialize_object_in_transaction;
+pub use materialization::{
+    enrich_objects_with_computed, enrich_objects_with_computed_query_snapshot,
+    preview_computed_definition, source_data_sha256,
+};
+use materialization::{evaluate_definitions, shared_definitions_conn, upsert_materialized};
+pub use query::{ComputedQuerySnapshot, resolve_computed_query_fields};
+pub(crate) use query::{computed_filter_predicate, object_cursor_sql_fields};
+#[cfg(test)]
+use query::{validate_computed_filter_count, validate_computed_query_count};
+pub(crate) use rebuild::{
+    enqueue_restored_computed_rebuilds, mark_recovered_computed_reindex_failed,
+};
+pub use rebuild::{
+    execute_computed_reindex_task, mark_computed_reindex_failed, request_class_rebuild,
+};
 
 const COMPUTED_CLASS_LOCK_NAMESPACE: i32 = 1_133_113;
 const REINDEX_PAYLOAD_TYPE: &str = "computed_fields";
@@ -170,637 +191,6 @@ pub async fn class_computation_state_for(
     })
     .await?
     .unwrap_or_else(|| ClassComputationState::ready_without_definitions(target_class_id)))
-}
-
-pub async fn resolve_computed_query_fields(
-    pool: &DbPool,
-    target_class_id: i32,
-    personal_owner_id: Option<i32>,
-    filters: &mut [ParsedQueryParam],
-    sorts: &mut [SortParam],
-) -> Result<ComputedSortSnapshot, ApiError> {
-    validate_computed_filter_count(
-        filters
-            .iter()
-            .filter(|filter| filter.field.computed_sort().is_some())
-            .count(),
-    )?;
-    let requested = filters
-        .iter()
-        .filter_map(|filter| filter.field.computed_sort())
-        .chain(sorts.iter().filter_map(|sort| sort.field.computed_sort()))
-        .map(|field| (field.scope(), field.key().to_string()))
-        .collect::<BTreeSet<_>>();
-    if requested.is_empty() {
-        return Err(ApiError::InternalServerError(
-            "Computed query resolution requires at least one computed field".to_string(),
-        ));
-    }
-    if sorts
-        .iter()
-        .any(|sort| sort.field.computed_sort().is_some())
-    {
-        validate_computed_sort_count(sorts.len())?;
-    }
-    if personal_owner_id.is_none()
-        && requested
-            .iter()
-            .any(|(scope, _)| *scope == ComputedFieldScope::Personal)
-    {
-        return Err(ApiError::BadRequest(
-            "Personal computed fields can only be filtered or sorted by their owning human user"
-                .to_string(),
-        ));
-    }
-
-    let (definitions, state) = with_transaction(pool, async |conn| -> Result<_, ApiError> {
-        acquire_computed_class_shared_lock(conn, target_class_id).await?;
-        use crate::schema::class_computation_state::dsl as state;
-        use crate::schema::computed_field_definitions::dsl as definition;
-        let mut query = definition::computed_field_definitions
-            .filter(definition::class_id.eq(target_class_id))
-            .into_boxed();
-        query = match personal_owner_id {
-            Some(owner_id) => query.filter(
-                definition::visibility
-                    .eq(COMPUTED_FIELD_VISIBILITY_SHARED)
-                    .or(definition::visibility
-                        .eq(COMPUTED_FIELD_VISIBILITY_PERSONAL)
-                        .and(definition::owner_user_id.eq(Some(owner_id)))),
-            ),
-            None => query.filter(definition::visibility.eq(COMPUTED_FIELD_VISIBILITY_SHARED)),
-        };
-        let definitions = query
-            .order(definition::id.asc())
-            .select(ComputedFieldDefinition::as_select())
-            .load::<ComputedFieldDefinition>(conn)
-            .await?;
-        let state = state::class_computation_state
-            .filter(state::class_id.eq(target_class_id))
-            .select(ClassComputationState::as_select())
-            .first(conn)
-            .await
-            .optional()?
-            .unwrap_or_else(|| ClassComputationState::ready_without_definitions(target_class_id));
-        Ok((definitions, state))
-    })
-    .await?;
-    for definition in definitions.iter().filter(|definition| definition.enabled) {
-        let _ = definition.evaluator_definition()?;
-    }
-    let definitions_by_key = definitions
-        .iter()
-        .filter(|definition| definition.enabled)
-        .map(|definition| {
-            let scope = if definition.is_shared() {
-                ComputedFieldScope::Shared
-            } else {
-                ComputedFieldScope::Personal
-            };
-            ((scope, definition.key.clone()), definition)
-        })
-        .collect::<HashMap<_, _>>();
-    let shared_scope_sql = computed_scope_sql(
-        definitions
-            .iter()
-            .filter(|definition| definition.is_shared() && definition.enabled),
-    )?;
-    let personal_scope_sql = computed_scope_sql(definitions.iter().filter(|definition| {
-        definition.enabled
-            && personal_owner_id.is_some_and(|owner_id| definition.is_personal_for(owner_id))
-    }))?;
-
-    let mut query_fields = HashMap::new();
-    for field in filters
-        .iter_mut()
-        .filter_map(|filter| filter.field.computed_sort_mut())
-        .chain(
-            sorts
-                .iter_mut()
-                .filter_map(|sort| sort.field.computed_sort_mut()),
-        )
-    {
-        let definition = definitions_by_key
-            .get(&(field.scope(), field.key().to_string()))
-            .ok_or_else(|| {
-                ApiError::BadRequest(format!(
-                    "Enabled {} computed field '{}' was not found for this class",
-                    field.scope().as_str(),
-                    field.key()
-                ))
-            })?;
-        let value_type =
-            crate::models::ComputedResultType::from_db(&definition.result_type)?.into();
-        let scope_sql = match field.scope() {
-            ComputedFieldScope::Shared => &shared_scope_sql,
-            ComputedFieldScope::Personal => &personal_scope_sql,
-        };
-        query_fields.insert(
-            (field.scope(), field.key().to_string()),
-            ResolvedComputedQueryField {
-                sql_expression: computed_sort_value_sql(
-                    definition,
-                    scope_sql,
-                    state.evaluation_revision,
-                )?,
-                value_type,
-            },
-        );
-        field.resolve(value_type);
-    }
-    Ok(ComputedSortSnapshot {
-        class_id: target_class_id,
-        definitions,
-        state,
-        query_fields,
-    })
-}
-
-fn validate_computed_filter_count(filter_count: usize) -> Result<(), ApiError> {
-    if filter_count > MAX_FILTERS_WITH_COMPUTED {
-        return Err(ApiError::BadRequest(format!(
-            "Computed filtering supports at most {MAX_FILTERS_WITH_COMPUTED} computed filter parameters per request"
-        )));
-    }
-    Ok(())
-}
-
-fn validate_computed_sort_count(sort_count: usize) -> Result<(), ApiError> {
-    if sort_count > MAX_SORT_FIELDS_WITH_COMPUTED {
-        return Err(ApiError::BadRequest(format!(
-            "Computed sorting supports at most {MAX_SORT_FIELDS_WITH_COMPUTED} explicit sort fields per request"
-        )));
-    }
-    Ok(())
-}
-
-#[derive(Debug, Clone)]
-pub struct ComputedSortSnapshot {
-    class_id: i32,
-    definitions: Vec<ComputedFieldDefinition>,
-    state: ClassComputationState,
-    query_fields: HashMap<(ComputedFieldScope, String), ResolvedComputedQueryField>,
-}
-
-#[derive(Debug, Clone)]
-struct ResolvedComputedQueryField {
-    sql_expression: String,
-    value_type: ComputedSortValueType,
-}
-
-impl ComputedSortSnapshot {
-    fn query_field(&self, field: &FilterField) -> Result<&ResolvedComputedQueryField, ApiError> {
-        let computed = field.computed_sort().ok_or_else(|| {
-            ApiError::InternalServerError(format!("Field '{field}' is not a computed field"))
-        })?;
-        self.query_fields
-            .get(&(computed.scope(), computed.key().to_string()))
-            .ok_or_else(|| {
-                ApiError::InternalServerError(format!(
-                    "Computed field '{}' was not resolved",
-                    computed.key()
-                ))
-            })
-    }
-}
-
-pub(crate) fn object_cursor_sql_fields(
-    sorts: &[SortParam],
-    snapshot: &ComputedSortSnapshot,
-) -> Result<Vec<OwnedCursorSqlField>, ApiError> {
-    sorts
-        .iter()
-        .map(|sort| {
-            if sort.field.computed_sort().is_none() {
-                return <HubuumObject as CursorSqlMapping>::sql_field(&sort.field).map(Into::into);
-            }
-            object_computed_sql_field(&sort.field, snapshot)
-        })
-        .collect()
-}
-
-fn object_computed_sql_field(
-    field: &FilterField,
-    snapshot: &ComputedSortSnapshot,
-) -> Result<OwnedCursorSqlField, ApiError> {
-    let resolved = snapshot.query_field(field)?;
-    let expression = &resolved.sql_expression;
-    let (expression, sql_type) = match resolved.value_type {
-        ComputedSortValueType::String => {
-            (format!("({expression} #>> '{{}}')"), CursorSqlType::String)
-        }
-        ComputedSortValueType::Number | ComputedSortValueType::Integer => (
-            format!("try_numeric({expression} #>> '{{}}')"),
-            CursorSqlType::Numeric,
-        ),
-        ComputedSortValueType::Boolean => (
-            format!("try_boolean({expression} #>> '{{}}')"),
-            CursorSqlType::Boolean,
-        ),
-        ComputedSortValueType::Object | ComputedSortValueType::Array => {
-            (expression.clone(), CursorSqlType::Json)
-        }
-    };
-    Ok(OwnedCursorSqlField {
-        expression,
-        sql_type,
-        nullable: true,
-    })
-}
-
-fn computed_scope_sql<'a>(
-    definitions: impl Iterator<Item = &'a ComputedFieldDefinition>,
-) -> Result<String, ApiError> {
-    let definitions = definitions
-        .map(|definition| {
-            serde_json::json!({
-                "key": definition.key,
-                "operation": definition.operation,
-                "result_type": definition.result_type,
-            })
-        })
-        .collect::<Vec<_>>();
-    Ok(serde_json::to_string(&definitions)?.replace('\'', "''"))
-}
-
-fn computed_sort_value_sql(
-    definition: &ComputedFieldDefinition,
-    scope_sql: &str,
-    evaluation_revision: i64,
-) -> Result<String, ApiError> {
-    let key = definition.key.replace('\'', "''");
-    let live_value = format!(
-        "NULLIF(\
-            hubuum_computed_evaluate_scope(\
-                hubuumobject.data, \
-                '{scope_sql}'::jsonb\
-            ) -> 'values' -> '{key}', \
-            'null'::jsonb\
-        )"
-    );
-    if !definition.is_shared() {
-        return Ok(live_value);
-    }
-    let cached_value = format!("sort_values.values -> '{key}'");
-    let cached_value_matches_type = match ComputedResultType::from_db(&definition.result_type)? {
-        ComputedResultType::String => format!("jsonb_typeof({cached_value}) = 'string'"),
-        ComputedResultType::Number => format!(
-            "jsonb_typeof({cached_value}) = 'number' \
-             AND hubuum_computed_numeric({cached_value}) IS NOT NULL"
-        ),
-        ComputedResultType::Integer => format!(
-            "jsonb_typeof({cached_value}) = 'number' \
-             AND hubuum_computed_numeric({cached_value}) IS NOT NULL \
-             AND trunc(hubuum_computed_numeric({cached_value})) = \
-                 hubuum_computed_numeric({cached_value})"
-        ),
-        ComputedResultType::Boolean => format!("jsonb_typeof({cached_value}) = 'boolean'"),
-        ComputedResultType::Object => format!("jsonb_typeof({cached_value}) = 'object'"),
-        ComputedResultType::Array => format!("jsonb_typeof({cached_value}) = 'array'"),
-    };
-    Ok(format!(
-        "(SELECT CASE \
-            WHEN sort_cache.present THEN sort_cache.value \
-            ELSE {live_value} \
-          END \
-          FROM (VALUES (TRUE)) AS sort_fallback(seed) \
-          LEFT JOIN LATERAL ( \
-            SELECT TRUE AS present, \
-                   NULLIF(sort_values.values -> '{key}', 'null'::jsonb) AS value \
-            FROM object_computed_data AS sort_values \
-            WHERE sort_values.object_id = hubuumobject.id \
-              AND sort_values.class_id = {} \
-              AND sort_values.evaluation_revision = {evaluation_revision} \
-              AND sort_values.source_data_sha256 = \
-                  hubuum_computed_source_sha256(hubuumobject.data) \
-              AND hubuum_computed_materialization_valid( \
-                  sort_values.values, \
-                  sort_values.errors, \
-                  '{scope_sql}'::jsonb \
-              ) \
-              AND jsonb_exists(sort_values.values, '{key}') \
-              AND ({cached_value} = 'null'::jsonb OR ({cached_value_matches_type})) \
-          ) AS sort_cache ON TRUE)",
-        definition.class_id
-    ))
-}
-
-pub(crate) fn computed_filter_predicate(
-    param: &ParsedQueryParam,
-    snapshot: &ComputedSortSnapshot,
-) -> Result<JsonSqlPredicate, ApiError> {
-    if param.value.contains('\0') {
-        return Err(ApiError::BadRequest(format!(
-            "Filter value for computed field '{}' contains a null character",
-            param.field
-        )));
-    }
-    let computed = param.field.computed_sort().ok_or_else(|| {
-        ApiError::InternalServerError(format!("Field '{}' is not computed", param.field))
-    })?;
-    let value_type = computed.value_type().ok_or_else(|| {
-        ApiError::InternalServerError(format!(
-            "Computed field '{}' has no resolved result type",
-            computed.key()
-        ))
-    })?;
-    let field = object_computed_sql_field(&param.field, snapshot)?;
-    let (operator, negated) = param.operator.op_and_neg();
-
-    let component = if operator == Operator::IsNull {
-        let should_be_null = param.value_as_boolean()? != negated;
-        SQLComponent {
-            sql: format!(
-                "{} IS {}NULL",
-                field.expression,
-                if should_be_null { "" } else { "NOT " }
-            ),
-            bind_variables: Vec::new(),
-        }
-    } else {
-        match value_type {
-            ComputedSortValueType::String => {
-                computed_string_filter(&field.expression, param, operator, negated)?
-            }
-            ComputedSortValueType::Number | ComputedSortValueType::Integer => {
-                computed_numeric_filter(&field.expression, param, operator, negated)?
-            }
-            ComputedSortValueType::Boolean => {
-                computed_boolean_filter(&field.expression, param, operator, negated)?
-            }
-            ComputedSortValueType::Object | ComputedSortValueType::Array => {
-                computed_json_filter(&field.expression, param, value_type, operator, negated)?
-            }
-        }
-    };
-    dynamic_sql_predicate(component)
-}
-
-fn computed_string_filter(
-    expression: &str,
-    param: &ParsedQueryParam,
-    operator: Operator,
-    negated: bool,
-) -> Result<SQLComponent, ApiError> {
-    let (sql, values) = match operator {
-        Operator::Equals => (format!("{expression} = ?"), vec![param.value.clone()]),
-        Operator::IEquals => (format!("{expression} ILIKE ?"), vec![param.value.clone()]),
-        Operator::Contains => (
-            format!("{expression} LIKE ?"),
-            vec![format!("%{}%", param.value)],
-        ),
-        Operator::IContains => (
-            format!("{expression} ILIKE ?"),
-            vec![format!("%{}%", param.value)],
-        ),
-        Operator::StartsWith => (
-            format!("{expression} LIKE ?"),
-            vec![format!("{}%", param.value)],
-        ),
-        Operator::IStartsWith => (
-            format!("{expression} ILIKE ?"),
-            vec![format!("{}%", param.value)],
-        ),
-        Operator::EndsWith => (
-            format!("{expression} LIKE ?"),
-            vec![format!("%{}", param.value)],
-        ),
-        Operator::IEndsWith => (
-            format!("{expression} ILIKE ?"),
-            vec![format!("%{}", param.value)],
-        ),
-        Operator::Like => (format!("{expression} LIKE ?"), vec![param.value.clone()]),
-        Operator::Regex => (format!("{expression} ~ ?"), vec![param.value.clone()]),
-        Operator::In => {
-            let values = exact_comma_separated_values(param, 50)?;
-            let placeholders = values.iter().map(|_| "?").collect::<Vec<_>>().join(", ");
-            (format!("{expression} IN ({placeholders})"), values)
-        }
-        _ => return Err(computed_operator_mismatch(param, "string")),
-    };
-    Ok(SQLComponent {
-        sql: maybe_negate(sql, negated),
-        bind_variables: values.into_iter().map(SQLValue::String).collect(),
-    })
-}
-
-fn computed_numeric_filter(
-    expression: &str,
-    param: &ParsedQueryParam,
-    operator: Operator,
-    negated: bool,
-) -> Result<SQLComponent, ApiError> {
-    let raw_values = trimmed_comma_separated_values(param, 50)?;
-    let values = raw_values
-        .iter()
-        .map(|value| {
-            hubuum_computed_fields::canonical_decimal_string(value).ok_or_else(|| {
-                ApiError::BadRequest(format!(
-                    "Invalid numeric value '{}' for computed field '{}'",
-                    value, param.field
-                ))
-            })
-        })
-        .collect::<Result<Vec<_>, _>>()?;
-    let sql = match operator {
-        Operator::Equals | Operator::In => {
-            let placeholders = values
-                .iter()
-                .map(|_| "?::numeric")
-                .collect::<Vec<_>>()
-                .join(", ");
-            format!("{expression} IN ({placeholders})")
-        }
-        Operator::Gt | Operator::Gte | Operator::Lt | Operator::Lte => {
-            require_computed_value_count(param, values.len(), 1)?;
-            let sql_operator = match operator {
-                Operator::Gt => ">",
-                Operator::Gte => ">=",
-                Operator::Lt => "<",
-                Operator::Lte => "<=",
-                _ => unreachable!(),
-            };
-            format!("{expression} {sql_operator} ?::numeric")
-        }
-        Operator::Between => {
-            require_computed_value_count(param, values.len(), 2)?;
-            format!("{expression} BETWEEN ?::numeric AND ?::numeric")
-        }
-        _ => return Err(computed_operator_mismatch(param, "numeric")),
-    };
-    Ok(SQLComponent {
-        sql: maybe_negate(sql, negated),
-        bind_variables: values.into_iter().map(SQLValue::String).collect(),
-    })
-}
-
-fn computed_boolean_filter(
-    expression: &str,
-    param: &ParsedQueryParam,
-    operator: Operator,
-    negated: bool,
-) -> Result<SQLComponent, ApiError> {
-    if operator != Operator::Equals {
-        return Err(computed_operator_mismatch(param, "boolean"));
-    }
-    Ok(SQLComponent {
-        sql: maybe_negate(format!("{expression} = ?"), negated),
-        bind_variables: vec![SQLValue::Boolean(param.value_as_boolean()?)],
-    })
-}
-
-fn computed_json_filter(
-    expression: &str,
-    param: &ParsedQueryParam,
-    value_type: ComputedSortValueType,
-    operator: Operator,
-    negated: bool,
-) -> Result<SQLComponent, ApiError> {
-    if operator == Operator::HasKey {
-        return Ok(SQLComponent {
-            sql: maybe_negate(format!("jsonb_exists({expression}, ?)"), negated),
-            bind_variables: vec![SQLValue::String(param.value.clone())],
-        });
-    }
-    if operator == Operator::ArrayLength && value_type == ComputedSortValueType::Array {
-        let length = param.value.parse::<i32>().map_err(|_| {
-            ApiError::BadRequest(format!(
-                "array_length requires an integer, got '{}'",
-                param.value
-            ))
-        })?;
-        if length < 0 {
-            return Err(ApiError::BadRequest(
-                "array_length requires a non-negative integer".to_string(),
-            ));
-        }
-        return Ok(SQLComponent {
-            sql: maybe_negate(format!("jsonb_array_length({expression}) = ?"), negated),
-            bind_variables: vec![SQLValue::Integer(length)],
-        });
-    }
-    if !matches!(operator, Operator::Equals | Operator::Contains) {
-        return Err(computed_operator_mismatch(param, value_type.as_str()));
-    }
-
-    let value: serde_json::Value = serde_json::from_str(&param.value).map_err(|error| {
-        ApiError::BadRequest(format!(
-            "Invalid JSON value for computed field '{}': {error}",
-            param.field
-        ))
-    })?;
-    let type_matches = matches!(
-        (value_type, &value),
-        (ComputedSortValueType::Object, serde_json::Value::Object(_))
-            | (ComputedSortValueType::Array, serde_json::Value::Array(_))
-    );
-    if !type_matches {
-        return Err(ApiError::BadRequest(format!(
-            "Filter value for computed field '{}' must be a JSON {}",
-            param.field,
-            value_type.as_str()
-        )));
-    }
-    validate_computed_filter_json(&value)?;
-    let json = serde_json::to_string(&value)?;
-    let sql_operator = if operator == Operator::Equals {
-        "="
-    } else {
-        "@>"
-    };
-    Ok(SQLComponent {
-        sql: maybe_negate(format!("{expression} {sql_operator} ?::jsonb"), negated),
-        bind_variables: vec![SQLValue::String(json)],
-    })
-}
-
-fn validate_computed_filter_json(value: &serde_json::Value) -> Result<(), ApiError> {
-    match crate::db::json::validate_postgres_jsonb_value(value) {
-        Ok(()) => Ok(()),
-        Err(crate::db::json::PostgresJsonbValidationError::UnsupportedValue) => {
-            Err(ApiError::BadRequest(
-                "Computed filter contains JSON that PostgreSQL JSONB cannot represent".to_string(),
-            ))
-        }
-        Err(crate::db::json::PostgresJsonbValidationError::NestingTooDeep) => {
-            Err(ApiError::BadRequest(format!(
-                "Computed filter JSON exceeds the maximum nesting depth of {}",
-                crate::db::json::MAX_POSTGRES_JSONB_NESTING_DEPTH
-            )))
-        }
-    }
-}
-
-fn exact_comma_separated_values(
-    param: &ParsedQueryParam,
-    maximum: usize,
-) -> Result<Vec<String>, ApiError> {
-    let values = param
-        .value
-        .split(',')
-        .map(str::to_string)
-        .collect::<Vec<_>>();
-    validate_comma_separated_values(param, maximum, &values)?;
-    Ok(values)
-}
-
-fn trimmed_comma_separated_values(
-    param: &ParsedQueryParam,
-    maximum: usize,
-) -> Result<Vec<String>, ApiError> {
-    let values = param
-        .value
-        .split(',')
-        .map(str::trim)
-        .map(str::to_string)
-        .collect::<Vec<_>>();
-    validate_comma_separated_values(param, maximum, &values)?;
-    Ok(values)
-}
-
-fn validate_comma_separated_values(
-    param: &ParsedQueryParam,
-    maximum: usize,
-    values: &[String],
-) -> Result<(), ApiError> {
-    if values.is_empty() || values.iter().any(String::is_empty) {
-        return Err(ApiError::BadRequest(format!(
-            "Filtering computed field '{}' requires a value",
-            param.field
-        )));
-    }
-    if values.len() > maximum {
-        return Err(ApiError::BadRequest(format!(
-            "Filtering computed field '{}' accepts at most {maximum} values",
-            param.field
-        )));
-    }
-    Ok(())
-}
-
-fn require_computed_value_count(
-    param: &ParsedQueryParam,
-    actual: usize,
-    expected: usize,
-) -> Result<(), ApiError> {
-    if actual != expected {
-        return Err(ApiError::OperatorMismatch(format!(
-            "Operator '{}' requires {expected} value(s) for field '{}'",
-            param.operator, param.field
-        )));
-    }
-    Ok(())
-}
-
-fn maybe_negate(sql: String, negated: bool) -> String {
-    if negated { format!("NOT ({sql})") } else { sql }
-}
-
-fn computed_operator_mismatch(param: &ParsedQueryParam, value_type: &str) -> ApiError {
-    ApiError::OperatorMismatch(format!(
-        "Operator '{}' is not applicable to computed field '{}' (type: {value_type})",
-        param.operator, param.field
-    ))
 }
 
 pub async fn list_shared_definitions(
@@ -1333,870 +723,6 @@ pub async fn delete_personal_definition(
     .await
 }
 
-pub async fn request_class_rebuild(
-    pool: &DbPool,
-    target_class_id: i32,
-    authorized_collection_id: i32,
-    actor_id: Option<i32>,
-) -> Result<ClassComputationState, ApiError> {
-    with_transaction(pool, async |conn| -> Result<_, ApiError> {
-        acquire_computed_class_exclusive_lock(conn, target_class_id).await?;
-        let _ = locked_class_record_in_collection(conn, target_class_id, authorized_collection_id)
-            .await?;
-        let state = ensure_computation_state(conn, target_class_id).await?;
-        if let Some(task_id_value) = state.active_task_id {
-            use crate::schema::tasks::dsl::{id, request_payload, status, tasks};
-            let active = tasks
-                .filter(id.eq(task_id_value))
-                .filter(status.eq_any([
-                    TaskStatus::Queued.as_str(),
-                    TaskStatus::Validating.as_str(),
-                    TaskStatus::Running.as_str(),
-                ]))
-                .select(request_payload)
-                .first::<Option<serde_json::Value>>(conn)
-                .await
-                .optional()?
-                .flatten()
-                .and_then(|value| serde_json::from_value::<ComputedReindexPayload>(value).ok())
-                .is_some_and(|payload| {
-                    payload.payload_type == REINDEX_PAYLOAD_TYPE
-                        && payload.class_id == target_class_id
-                        && payload.target_revision == state.evaluation_revision
-                });
-            if active {
-                return Ok(state);
-            }
-        }
-        enqueue_rebuild(conn, target_class_id, state.evaluation_revision, actor_id).await
-    })
-    .await
-}
-
-pub(crate) async fn enqueue_restored_computed_rebuilds(
-    conn: &mut DbConnection,
-) -> Result<(), ApiError> {
-    use crate::schema::computed_field_definitions::dsl::{
-        class_id, computed_field_definitions, visibility,
-    };
-    let class_ids = computed_field_definitions
-        .filter(visibility.eq(COMPUTED_FIELD_VISIBILITY_SHARED))
-        .select(class_id)
-        .distinct()
-        .order(class_id.asc())
-        .load::<i32>(conn)
-        .await?;
-    for target_class_id in class_ids {
-        acquire_computed_class_exclusive_lock(conn, target_class_id).await?;
-        advance_revision_and_enqueue(conn, target_class_id, None).await?;
-    }
-    Ok(())
-}
-
-fn shared_definitions_from_rows(
-    definitions: &[ComputedFieldDefinition],
-) -> Result<Vec<hubuum_computed_fields::Definition>, ApiError> {
-    definitions
-        .iter()
-        .filter(|definition| definition.enabled)
-        .map(ComputedFieldDefinition::evaluator_definition)
-        .collect()
-}
-
-fn evaluate_definitions(
-    data: &serde_json::Value,
-    definitions: &[ComputedFieldDefinition],
-    maximum: usize,
-    scope: &'static str,
-) -> Result<EvaluationResult, ApiError> {
-    let definitions = shared_definitions_from_rows(definitions)?;
-    let result =
-        evaluate(data, &definitions, maximum, EvaluationLimits::standard()).map_err(|error| {
-            ApiError::InternalServerError(format!("Computed-field evaluation failed: {error}"))
-        })?;
-    crate::observability::metrics::computed_evaluation(scope, &result);
-    Ok(result)
-}
-
-pub fn preview_computed_definition(
-    data: &serde_json::Value,
-    request: &ComputedFieldDefinitionRequest,
-) -> Result<crate::models::ComputedFieldPreviewResponse, ApiError> {
-    let definition = request.validate()?;
-    let result =
-        evaluate(data, &[definition], 1, EvaluationLimits::standard()).map_err(|error| {
-            ApiError::InternalServerError(format!("Computed-field preview failed: {error}"))
-        })?;
-    crate::observability::metrics::computed_evaluation("preview", &result);
-    Ok(crate::models::ComputedFieldPreviewResponse {
-        value: result
-            .values
-            .get(&request.key)
-            .cloned()
-            .unwrap_or(serde_json::Value::Null),
-        error: result
-            .errors
-            .get(&request.key)
-            .cloned()
-            .map(ComputedFieldErrorResponse::from),
-    })
-}
-
-fn canonical_json(value: &serde_json::Value, output: &mut String) -> Result<(), ApiError> {
-    match value {
-        serde_json::Value::Object(values) => {
-            output.push('{');
-            let mut keys = values.keys().collect::<Vec<_>>();
-            keys.sort_unstable();
-            for (index, key) in keys.into_iter().enumerate() {
-                if index > 0 {
-                    output.push(',');
-                }
-                output.push_str(&serde_json::to_string(key)?);
-                output.push(':');
-                canonical_json(&values[key], output)?;
-            }
-            output.push('}');
-        }
-        serde_json::Value::Array(values) => {
-            output.push('[');
-            for (index, value) in values.iter().enumerate() {
-                if index > 0 {
-                    output.push(',');
-                }
-                canonical_json(value, output)?;
-            }
-            output.push(']');
-        }
-        _ => output.push_str(&serde_json::to_string(value)?),
-    }
-    Ok(())
-}
-
-pub fn source_data_sha256(data: &serde_json::Value) -> Result<String, ApiError> {
-    let mut canonical = String::new();
-    canonical_json(data, &mut canonical)?;
-    Ok(Sha256::digest(canonical.as_bytes())
-        .iter()
-        .map(|byte| format!("{byte:02x}"))
-        .collect())
-}
-
-async fn upsert_materialized(
-    conn: &mut DbConnection,
-    object: &HubuumObject,
-    revision: i64,
-    result: EvaluationResult,
-) -> Result<(), ApiError> {
-    use crate::schema::object_computed_data::dsl::{
-        class_id, computed_at, errors, evaluation_revision, object_computed_data, object_id,
-        source_data_sha256 as stored_hash, values,
-    };
-    let input = NewObjectComputedData {
-        object_id: object.id,
-        class_id: object.hubuum_class_id,
-        evaluation_revision: revision,
-        source_data_sha256: source_data_sha256(&object.data)?,
-        values: serde_json::to_value(result.values)?,
-        errors: serde_json::to_value(result.errors)?,
-    };
-    diesel::insert_into(object_computed_data)
-        .values(&input)
-        .on_conflict(object_id)
-        .do_update()
-        .set((
-            class_id.eq(input.class_id),
-            evaluation_revision.eq(input.evaluation_revision),
-            stored_hash.eq(&input.source_data_sha256),
-            values.eq(&input.values),
-            errors.eq(&input.errors),
-            computed_at.eq(diesel::dsl::now),
-        ))
-        .execute(conn)
-        .await?;
-    Ok(())
-}
-
-async fn shared_definitions_conn(
-    conn: &mut DbConnection,
-    target_class_id: i32,
-) -> Result<Vec<ComputedFieldDefinition>, ApiError> {
-    use crate::schema::computed_field_definitions::dsl::{
-        class_id, computed_field_definitions, id, visibility,
-    };
-    Ok(computed_field_definitions
-        .filter(class_id.eq(target_class_id))
-        .filter(visibility.eq(COMPUTED_FIELD_VISIBILITY_SHARED))
-        .order(id.asc())
-        .select(ComputedFieldDefinition::as_select())
-        .load(conn)
-        .await?)
-}
-
-/// Materialize one canonical object inside the caller's write transaction.
-pub(crate) async fn materialize_object_in_transaction(
-    conn: &mut DbConnection,
-    object: &HubuumObject,
-) -> Result<(), ApiError> {
-    acquire_computed_class_shared_lock(conn, object.hubuum_class_id).await?;
-    let definitions = shared_definitions_conn(conn, object.hubuum_class_id).await?;
-    if definitions.is_empty() {
-        use crate::schema::object_computed_data::dsl::{object_computed_data, object_id};
-        diesel::delete(object_computed_data.filter(object_id.eq(object.id)))
-            .execute(conn)
-            .await?;
-        return Ok(());
-    }
-    let state = ensure_computation_state(conn, object.hubuum_class_id).await?;
-    let result =
-        evaluate_definitions(&object.data, &definitions, MAX_SHARED_DEFINITIONS, "shared")?;
-    upsert_materialized(conn, object, state.evaluation_revision, result).await
-}
-
-type ResponseEvaluationMaps = (
-    BTreeMap<String, serde_json::Value>,
-    BTreeMap<String, ComputedFieldErrorResponse>,
-);
-
-#[derive(Clone, Copy)]
-enum MaterializationRepair {
-    Apply,
-    Defer,
-}
-
-fn evaluation_maps(result: EvaluationResult) -> ResponseEvaluationMaps {
-    (
-        result.values,
-        result
-            .errors
-            .into_iter()
-            .map(|(key, error)| (key, error.into()))
-            .collect(),
-    )
-}
-
-fn valid_stored_evaluation_maps(
-    row: &ObjectComputedData,
-    definitions: &[ComputedFieldDefinition],
-) -> Result<Option<ResponseEvaluationMaps>, ApiError> {
-    let Ok(values) =
-        serde_json::from_value::<BTreeMap<String, serde_json::Value>>(row.values.clone())
-    else {
-        return Ok(None);
-    };
-    let Ok(errors) = serde_json::from_value::<BTreeMap<String, hubuum_computed_fields::FieldError>>(
-        row.errors.clone(),
-    ) else {
-        return Ok(None);
-    };
-    let enabled = definitions
-        .iter()
-        .filter(|definition| definition.enabled)
-        .collect::<Vec<_>>();
-    if values.len() != enabled.len() {
-        return Ok(None);
-    }
-    for definition in &enabled {
-        let Some(value) = values.get(&definition.key) else {
-            return Ok(None);
-        };
-        let result_type = ComputedResultType::from_db(&definition.result_type)?;
-        if !computed_value_matches_result_type(value, result_type) {
-            return Ok(None);
-        }
-    }
-    let enabled_keys = enabled
-        .iter()
-        .map(|definition| definition.key.as_str())
-        .collect::<BTreeSet<_>>();
-    if errors.iter().any(|(key, _)| {
-        !enabled_keys.contains(key.as_str())
-            || !values.get(key).is_some_and(serde_json::Value::is_null)
-    }) {
-        return Ok(None);
-    }
-
-    Ok(Some((
-        values,
-        errors
-            .into_iter()
-            .map(|(key, error)| (key, error.into()))
-            .collect(),
-    )))
-}
-
-fn computed_value_matches_result_type(
-    value: &serde_json::Value,
-    result_type: ComputedResultType,
-) -> bool {
-    if value.is_null() {
-        return true;
-    }
-    match result_type {
-        ComputedResultType::String => value.is_string(),
-        ComputedResultType::Number => value.as_number().is_some_and(|number| {
-            hubuum_computed_fields::canonical_decimal_string(&number.to_string()).is_some()
-        }),
-        ComputedResultType::Integer => value.as_number().is_some_and(|number| {
-            hubuum_computed_fields::canonical_integer_string(&number.to_string()).is_some()
-        }),
-        ComputedResultType::Boolean => value.is_boolean(),
-        ComputedResultType::Object => value.is_object(),
-        ComputedResultType::Array => value.is_array(),
-    }
-}
-
-pub async fn enrich_objects_with_computed(
-    pool: &DbPool,
-    objects: Vec<HubuumObject>,
-    personal_owner_id: Option<i32>,
-) -> Result<Vec<HubuumObjectComputedResponse>, ApiError> {
-    if objects.is_empty() {
-        return Ok(Vec::new());
-    }
-    let class_ids = objects
-        .iter()
-        .map(|object| object.hubuum_class_id)
-        .collect::<BTreeSet<_>>();
-    let object_ids = objects.iter().map(|object| object.id).collect::<Vec<_>>();
-    let class_id_values = class_ids.iter().copied().collect::<Vec<_>>();
-
-    let (definitions, states, materialized) =
-        with_transaction(pool, async |conn| -> Result<_, ApiError> {
-            for target_class_id in &class_ids {
-                acquire_computed_class_shared_lock(conn, *target_class_id).await?;
-            }
-            use crate::schema::class_computation_state::dsl as state;
-            use crate::schema::computed_field_definitions::dsl as definition;
-            use crate::schema::object_computed_data::dsl as computed;
-            let mut definition_query = definition::computed_field_definitions
-                .filter(definition::class_id.eq_any(&class_id_values))
-                .into_boxed();
-            definition_query = match personal_owner_id {
-                Some(owner_id) => definition_query.filter(
-                    definition::visibility
-                        .eq(COMPUTED_FIELD_VISIBILITY_SHARED)
-                        .or(definition::visibility
-                            .eq(COMPUTED_FIELD_VISIBILITY_PERSONAL)
-                            .and(definition::owner_user_id.eq(Some(owner_id)))),
-                ),
-                None => definition_query
-                    .filter(definition::visibility.eq(COMPUTED_FIELD_VISIBILITY_SHARED)),
-            };
-            let definitions = definition_query
-                .order((definition::class_id.asc(), definition::id.asc()))
-                .select(ComputedFieldDefinition::as_select())
-                .load::<ComputedFieldDefinition>(conn)
-                .await?;
-            let states = state::class_computation_state
-                .filter(state::class_id.eq_any(&class_id_values))
-                .select(ClassComputationState::as_select())
-                .load::<ClassComputationState>(conn)
-                .await?;
-            let materialized = computed::object_computed_data
-                .filter(computed::object_id.eq_any(&object_ids))
-                .select(ObjectComputedData::as_select())
-                .load::<ObjectComputedData>(conn)
-                .await?;
-            Ok((definitions, states, materialized))
-        })
-        .await?;
-
-    enrich_objects_from_snapshot(
-        pool,
-        objects,
-        personal_owner_id,
-        definitions,
-        states,
-        materialized,
-        MaterializationRepair::Apply,
-    )
-    .await
-}
-
-pub async fn enrich_objects_with_computed_sort_snapshot(
-    pool: &DbPool,
-    objects: Vec<HubuumObject>,
-    personal_owner_id: Option<i32>,
-    snapshot: &ComputedSortSnapshot,
-) -> Result<Vec<HubuumObjectComputedResponse>, ApiError> {
-    if objects.is_empty() {
-        return Ok(Vec::new());
-    }
-    if objects
-        .iter()
-        .any(|object| object.hubuum_class_id != snapshot.class_id)
-    {
-        return Err(ApiError::InternalServerError(
-            "Computed sort snapshot cannot enrich objects from another class".to_string(),
-        ));
-    }
-    let object_ids = objects.iter().map(|object| object.id).collect::<Vec<_>>();
-    let materialized = with_connection(pool, async |conn| {
-        use crate::schema::object_computed_data::dsl as computed;
-        computed::object_computed_data
-            .filter(computed::object_id.eq_any(&object_ids))
-            .select(ObjectComputedData::as_select())
-            .load::<ObjectComputedData>(conn)
-            .await
-    })
-    .await?;
-
-    enrich_objects_from_snapshot(
-        pool,
-        objects,
-        personal_owner_id,
-        snapshot.definitions.clone(),
-        vec![snapshot.state.clone()],
-        materialized,
-        MaterializationRepair::Defer,
-    )
-    .await
-}
-
-async fn enrich_objects_from_snapshot(
-    pool: &DbPool,
-    objects: Vec<HubuumObject>,
-    personal_owner_id: Option<i32>,
-    definitions: Vec<ComputedFieldDefinition>,
-    states: Vec<ClassComputationState>,
-    materialized: Vec<ObjectComputedData>,
-    repair: MaterializationRepair,
-) -> Result<Vec<HubuumObjectComputedResponse>, ApiError> {
-    let mut shared_by_class: HashMap<i32, Vec<ComputedFieldDefinition>> = HashMap::new();
-    let mut personal_by_class: HashMap<i32, Vec<ComputedFieldDefinition>> = HashMap::new();
-    for definition in definitions {
-        if definition.is_shared() {
-            shared_by_class
-                .entry(definition.class_id)
-                .or_default()
-                .push(definition);
-        } else if personal_owner_id.is_some_and(|owner| definition.is_personal_for(owner)) {
-            personal_by_class
-                .entry(definition.class_id)
-                .or_default()
-                .push(definition);
-        }
-    }
-    let states = states
-        .into_iter()
-        .map(|state| (state.class_id, state))
-        .collect::<HashMap<_, _>>();
-    let materialized = materialized
-        .into_iter()
-        .map(|row| (row.object_id, row))
-        .collect::<HashMap<_, _>>();
-
-    let mut stale_objects = Vec::new();
-    let mut enriched = Vec::with_capacity(objects.len());
-    for object in objects {
-        let state = states
-            .get(&object.hubuum_class_id)
-            .cloned()
-            .unwrap_or_else(|| {
-                ClassComputationState::ready_without_definitions(object.hubuum_class_id)
-            });
-        let definitions = shared_by_class
-            .get(&object.hubuum_class_id)
-            .map(Vec::as_slice)
-            .unwrap_or(&[]);
-        let hash = source_data_sha256(&object.data)?;
-        let stored = materialized.get(&object.id);
-        let has_enabled_definitions = definitions.iter().any(|definition| definition.enabled);
-        let stored_maps = match stored {
-            Some(row)
-                if row.class_id == object.hubuum_class_id
-                    && row.evaluation_revision == state.evaluation_revision
-                    && row.source_data_sha256 == hash =>
-            {
-                valid_stored_evaluation_maps(row, definitions)?
-            }
-            _ => None,
-        };
-        let fresh = !has_enabled_definitions || stored_maps.is_some();
-        let (shared_values, shared_errors) = if !has_enabled_definitions {
-            (BTreeMap::new(), BTreeMap::new())
-        } else if let Some(maps) = stored_maps {
-            maps
-        } else {
-            if matches!(repair, MaterializationRepair::Apply) {
-                stale_objects.push(object.clone());
-            }
-            crate::observability::metrics::computed_live_fallback();
-            evaluation_maps(evaluate_definitions(
-                &object.data,
-                definitions,
-                MAX_SHARED_DEFINITIONS,
-                "shared",
-            )?)
-        };
-
-        let personal = if personal_owner_id.is_some() {
-            let definitions = personal_by_class
-                .get(&object.hubuum_class_id)
-                .map(Vec::as_slice)
-                .unwrap_or(&[]);
-            let result = evaluate_definitions(
-                &object.data,
-                definitions,
-                MAX_PERSONAL_DEFINITIONS,
-                "personal",
-            )?;
-            let (values, errors) = evaluation_maps(result);
-            Some(ComputedScopeResponse { values, errors })
-        } else {
-            None
-        };
-        enriched.push(HubuumObjectComputedResponse {
-            object,
-            computed: ComputedObjectScopesResponse {
-                shared: SharedComputedScopeResponse {
-                    revision: state.evaluation_revision,
-                    materialization_stale: !fresh,
-                    values: shared_values,
-                    errors: shared_errors,
-                },
-                personal,
-            },
-        });
-    }
-
-    if !stale_objects.is_empty() {
-        match repair_stale_materializations(pool, stale_objects).await {
-            Ok(()) => crate::observability::metrics::computed_read_repair("success"),
-            Err(error) => {
-                crate::observability::metrics::computed_read_repair("failure");
-                warn!(message = "Computed-field read repair failed", error = %error);
-            }
-        }
-    }
-    Ok(enriched)
-}
-
-async fn repair_stale_materializations(
-    pool: &DbPool,
-    mut stale_objects: Vec<HubuumObject>,
-) -> Result<(), ApiError> {
-    stale_objects.sort_by_key(|object| object.id);
-    let object_ids = stale_objects
-        .iter()
-        .map(|object| object.id)
-        .collect::<Vec<_>>();
-    with_transaction(pool, async |conn| -> Result<_, ApiError> {
-        use crate::schema::hubuumobject::dsl::{hubuumobject, id};
-        let current_objects = hubuumobject
-            .filter(id.eq_any(&object_ids))
-            .order(id.asc())
-            .for_update()
-            .load::<HubuumObject>(conn)
-            .await?;
-
-        // Object writes lock their row before taking the class advisory lock.
-        // Repairs and rebuild batches use the same order so an exclusive
-        // definition update cannot form a row/advisory-lock cycle.
-        let class_ids = current_objects
-            .iter()
-            .map(|object| object.hubuum_class_id)
-            .collect::<BTreeSet<_>>();
-        for class_id in &class_ids {
-            acquire_computed_class_shared_lock(conn, *class_id).await?;
-        }
-
-        let class_id_values = class_ids.iter().copied().collect::<Vec<_>>();
-        use crate::schema::computed_field_definitions::dsl as definition;
-        let definitions = definition::computed_field_definitions
-            .filter(definition::class_id.eq_any(&class_id_values))
-            .filter(definition::visibility.eq(COMPUTED_FIELD_VISIBILITY_SHARED))
-            .order((definition::class_id.asc(), definition::id.asc()))
-            .select(ComputedFieldDefinition::as_select())
-            .load::<ComputedFieldDefinition>(conn)
-            .await?;
-        let mut definitions_by_class: HashMap<i32, Vec<ComputedFieldDefinition>> = HashMap::new();
-        for definition in definitions {
-            definitions_by_class
-                .entry(definition.class_id)
-                .or_default()
-                .push(definition);
-        }
-        let mut revisions = HashMap::new();
-        for class_id in class_ids {
-            let state = ensure_computation_state(conn, class_id).await?;
-            revisions.insert(class_id, state.evaluation_revision);
-        }
-
-        let mut without_definitions = Vec::new();
-        for object in current_objects {
-            let definitions = definitions_by_class
-                .get(&object.hubuum_class_id)
-                .map(Vec::as_slice)
-                .unwrap_or(&[]);
-            if definitions.is_empty() {
-                without_definitions.push(object.id);
-                continue;
-            }
-            let result =
-                evaluate_definitions(&object.data, definitions, MAX_SHARED_DEFINITIONS, "shared")?;
-            upsert_materialized(conn, &object, revisions[&object.hubuum_class_id], result).await?;
-        }
-        if !without_definitions.is_empty() {
-            use crate::schema::object_computed_data::dsl::{object_computed_data, object_id};
-            diesel::delete(object_computed_data.filter(object_id.eq_any(without_definitions)))
-                .execute(conn)
-                .await?;
-        }
-        Ok(())
-    })
-    .await
-}
-
-enum ReindexBatch {
-    Superseded,
-    Rows { last_id: i32, count: i32 },
-    Complete,
-}
-
-async fn process_reindex_batch(
-    pool: &DbPool,
-    task: &TaskRecord,
-    payload: &ComputedReindexPayload,
-    cursor: i32,
-) -> Result<ReindexBatch, ApiError> {
-    with_transaction(pool, async |conn| -> Result<_, ApiError> {
-        use crate::schema::hubuumobject::dsl::{hubuum_class_id, hubuumobject, id};
-        let objects = hubuumobject
-            .filter(hubuum_class_id.eq(payload.class_id))
-            .filter(id.gt(cursor))
-            .filter(id.le(payload.object_upper_bound))
-            .order(id.asc())
-            .limit(reindex_batch_size())
-            .for_update()
-            .load::<HubuumObject>(conn)
-            .await?;
-        acquire_computed_class_shared_lock(conn, payload.class_id).await?;
-        let state = ensure_computation_state(conn, payload.class_id).await?;
-        if state.evaluation_revision != payload.target_revision
-            || state.active_task_id != Some(task.id)
-        {
-            return Ok(ReindexBatch::Superseded);
-        }
-        if objects.is_empty() {
-            return Ok(ReindexBatch::Complete);
-        }
-        let definitions = shared_definitions_conn(conn, payload.class_id).await?;
-        for object in &objects {
-            if definitions.is_empty() {
-                use crate::schema::object_computed_data::dsl::{object_computed_data, object_id};
-                diesel::delete(object_computed_data.filter(object_id.eq(object.id)))
-                    .execute(conn)
-                    .await?;
-            } else {
-                let result = evaluate_definitions(
-                    &object.data,
-                    &definitions,
-                    MAX_SHARED_DEFINITIONS,
-                    "shared",
-                )?;
-                upsert_materialized(conn, object, payload.target_revision, result).await?;
-            }
-        }
-        Ok(ReindexBatch::Rows {
-            last_id: objects.last().expect("non-empty batch").id,
-            count: objects.len() as i32,
-        })
-    })
-    .await
-}
-
-pub async fn execute_computed_reindex_task(
-    pool: &DbPool,
-    task: &TaskRecord,
-) -> Result<(), ApiError> {
-    let started = Instant::now();
-    let payload: ComputedReindexPayload = serde_json::from_value(
-        task.request_payload
-            .clone()
-            .ok_or_else(|| ApiError::BadRequest("Reindex task payload is missing".to_string()))?,
-    )?;
-    payload.validate()?;
-    let mut cursor = 0;
-    let mut processed = 0;
-    loop {
-        match process_reindex_batch(pool, task, &payload, cursor).await? {
-            ReindexBatch::Superseded => {
-                task.finalize_terminal(
-                    pool,
-                    TaskStateUpdate {
-                        status: TaskStatus::Cancelled,
-                        summary: Some("Computed-field rebuild superseded".to_string()),
-                        processed_items: processed,
-                        success_items: processed,
-                        failed_items: 0,
-                        started_at: task.started_at,
-                        finished_at: None,
-                    },
-                    NewTaskEventRecord {
-                        task_id: task.id,
-                        event_type: TaskStatus::Cancelled.as_str().to_string(),
-                        message: "Computed-field rebuild superseded".to_string(),
-                        data: None,
-                    },
-                )
-                .await?;
-                crate::observability::metrics::computed_rebuild_finished(
-                    "cancelled",
-                    started.elapsed(),
-                );
-                return Ok(());
-            }
-            ReindexBatch::Rows { last_id, count } => {
-                crate::observability::metrics::computed_rebuild_batch(count as usize);
-                cursor = last_id;
-                processed = processed.saturating_add(count);
-                task.update_state(
-                    pool,
-                    TaskStateUpdate {
-                        status: TaskStatus::Running,
-                        summary: Some(format!(
-                            "Rebuilt {processed} of {} objects",
-                            task.total_items
-                        )),
-                        processed_items: processed,
-                        success_items: processed,
-                        failed_items: 0,
-                        started_at: task.started_at,
-                        finished_at: None,
-                    },
-                )
-                .await?;
-            }
-            ReindexBatch::Complete => {
-                crate::observability::metrics::computed_rebuild_batch(0);
-                break;
-            }
-        }
-    }
-
-    let ready = with_transaction(pool, async |conn| -> Result<bool, ApiError> {
-        acquire_computed_class_shared_lock(conn, payload.class_id).await?;
-        use crate::schema::class_computation_state::dsl::{
-            active_task_id, class_computation_state, class_id, evaluation_revision, last_error,
-            rebuild_status, updated_at,
-        };
-        let changed = diesel::update(
-            class_computation_state
-                .filter(class_id.eq(payload.class_id))
-                .filter(evaluation_revision.eq(payload.target_revision))
-                .filter(active_task_id.eq(Some(task.id))),
-        )
-        .set((
-            rebuild_status.eq("ready"),
-            active_task_id.eq::<Option<i32>>(None),
-            last_error.eq::<Option<String>>(None),
-            updated_at.eq(diesel::dsl::now),
-        ))
-        .execute(conn)
-        .await?;
-        Ok(changed == 1)
-    })
-    .await?;
-    let (status, summary) = if ready {
-        (
-            TaskStatus::Succeeded,
-            format!("Computed-field rebuild completed for {processed} objects"),
-        )
-    } else {
-        (
-            TaskStatus::Cancelled,
-            "Computed-field rebuild superseded before completion".to_string(),
-        )
-    };
-    task.finalize_terminal(
-        pool,
-        TaskStateUpdate {
-            status,
-            summary: Some(summary.clone()),
-            processed_items: processed,
-            success_items: processed,
-            failed_items: 0,
-            started_at: task.started_at,
-            finished_at: None,
-        },
-        NewTaskEventRecord {
-            task_id: task.id,
-            event_type: status.as_str().to_string(),
-            message: summary,
-            data: None,
-        },
-    )
-    .await?;
-    crate::observability::metrics::computed_rebuild_finished(status.as_str(), started.elapsed());
-    info!(
-        message = "Computed-field rebuild finished",
-        task_id = task.id,
-        class_id = payload.class_id,
-        revision = payload.target_revision,
-        processed
-    );
-    Ok(())
-}
-
-pub async fn mark_computed_reindex_failed(
-    pool: &DbPool,
-    task: &TaskRecord,
-    stored_error: &str,
-) -> Result<(), ApiError> {
-    let Some(payload) = task
-        .request_payload
-        .clone()
-        .and_then(|value| serde_json::from_value::<ComputedReindexPayload>(value).ok())
-    else {
-        return Ok(());
-    };
-    with_connection(pool, async |conn| {
-        use crate::schema::class_computation_state::dsl::{
-            active_task_id, class_computation_state, class_id, evaluation_revision, last_error,
-            rebuild_status, updated_at,
-        };
-        diesel::update(
-            class_computation_state
-                .filter(class_id.eq(payload.class_id))
-                .filter(evaluation_revision.eq(payload.target_revision))
-                .filter(active_task_id.eq(Some(task.id))),
-        )
-        .set((
-            rebuild_status.eq("failed"),
-            active_task_id.eq::<Option<i32>>(None),
-            last_error.eq(Some(stored_error.chars().take(512).collect::<String>())),
-            updated_at.eq(diesel::dsl::now),
-        ))
-        .execute(conn)
-        .await
-    })
-    .await?;
-    crate::observability::metrics::computed_rebuild_finished("failed", std::time::Duration::ZERO);
-    Ok(())
-}
-
-pub(crate) async fn mark_recovered_computed_reindex_failed(
-    conn: &mut DbConnection,
-    task_id: i32,
-    stored_error: &str,
-) -> Result<(), ApiError> {
-    use crate::schema::class_computation_state::dsl::{
-        active_task_id, class_computation_state, last_error, rebuild_status, updated_at,
-    };
-    diesel::update(class_computation_state.filter(active_task_id.eq(Some(task_id))))
-        .set((
-            rebuild_status.eq("failed"),
-            active_task_id.eq::<Option<i32>>(None),
-            last_error.eq(Some(stored_error.chars().take(512).collect::<String>())),
-            updated_at.eq(diesel::dsl::now),
-        ))
-        .execute(conn)
-        .await?;
-    Ok(())
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
@@ -2224,12 +750,12 @@ mod tests {
     }
 
     #[test]
-    fn computed_sort_request_size_is_bounded_before_sql_resolution() {
+    fn computed_query_request_size_is_bounded_before_sql_resolution() {
         let sorts = parse_query_parameter("sort=computed.shared.first,computed.shared.second,name")
             .unwrap()
             .sort;
 
-        let error = validate_computed_sort_count(sorts.len()).unwrap_err();
+        let error = validate_computed_query_count(sorts.len()).unwrap_err();
 
         assert_eq!(
             error.to_string(),
@@ -2247,7 +773,7 @@ mod tests {
         let computed_filter_count = query
             .filters
             .iter()
-            .filter(|filter| filter.field.computed_sort().is_some())
+            .filter(|filter| filter.field.computed_query().is_some())
             .count();
 
         let error = validate_computed_filter_count(computed_filter_count).unwrap_err();
