+++ src/models/computed_field.rs
@@ -10,7 +10,7 @@ use utoipa::ToSchema;
 use crate::db::prelude::*;
 use crate::errors::ApiError;
 use crate::models::HubuumObject;
-use crate::models::search::{ComputedSortValueType, FilterField, SortParam};
+use crate::models::search::{ComputedQueryValueType, FilterField, SortParam};
 use crate::pagination::{
     CursorPaginated, CursorSqlField, CursorSqlMapping, CursorSqlType, CursorValue,
 };
@@ -76,7 +76,7 @@ impl From<ComputedResultType> for ResultType {
     }
 }
 
-impl From<ComputedResultType> for ComputedSortValueType {
+impl From<ComputedResultType> for ComputedQueryValueType {
     fn from(value: ComputedResultType) -> Self {
         match value {
             ComputedResultType::String => Self::String,
@@ -535,14 +535,14 @@ pub struct HubuumObjectComputedResponse {
 
 impl CursorPaginated for HubuumObjectComputedResponse {
     fn supports_sort(field: &FilterField) -> bool {
-        field.computed_sort().is_some() || HubuumObject::supports_sort(field)
+        field.computed_query().is_some() || HubuumObject::supports_sort(field)
     }
 
     fn cursor_value(&self, field: &FilterField) -> Result<CursorValue, ApiError> {
-        let Some(computed_sort) = field.computed_sort() else {
+        let Some(computed_query) = field.computed_query() else {
             return self.object.cursor_value(field);
         };
-        let values = match computed_sort.scope() {
+        let values = match computed_query.scope() {
             crate::models::search::ComputedFieldScope::Shared => &self.computed.shared.values,
             crate::models::search::ComputedFieldScope::Personal => {
                 &self
@@ -558,41 +558,41 @@ impl CursorPaginated for HubuumObjectComputedResponse {
                     .values
             }
         };
-        let value = values.get(computed_sort.key()).ok_or_else(|| {
+        let value = values.get(computed_query.key()).ok_or_else(|| {
             ApiError::InternalServerError(format!(
                 "Computed sort field '{}' is missing from the enriched object",
-                computed_sort.key()
+                computed_query.key()
             ))
         })?;
         if value.is_null() {
             return Ok(CursorValue::Null);
         }
-        let value_type = computed_sort.value_type().ok_or_else(|| {
+        let value_type = computed_query.value_type().ok_or_else(|| {
             ApiError::InternalServerError(format!(
                 "Computed sort field '{}' was not resolved",
-                computed_sort.key()
+                computed_query.key()
             ))
         })?;
         match value_type {
-            ComputedSortValueType::String => value
+            ComputedQueryValueType::String => value
                 .as_str()
                 .map(|value| CursorValue::String(value.to_string())),
-            ComputedSortValueType::Number | ComputedSortValueType::Integer => value
+            ComputedQueryValueType::Number | ComputedQueryValueType::Integer => value
                 .is_number()
                 .then(|| CursorValue::Decimal(value.to_string())),
-            ComputedSortValueType::Boolean => value.as_bool().map(CursorValue::Boolean),
-            ComputedSortValueType::Object if value.is_object() => {
+            ComputedQueryValueType::Boolean => value.as_bool().map(CursorValue::Boolean),
+            ComputedQueryValueType::Object if value.is_object() => {
                 Some(CursorValue::Json(value.clone()))
             }
-            ComputedSortValueType::Array if value.is_array() => {
+            ComputedQueryValueType::Array if value.is_array() => {
                 Some(CursorValue::Json(value.clone()))
             }
-            ComputedSortValueType::Object | ComputedSortValueType::Array => None,
+            ComputedQueryValueType::Object | ComputedQueryValueType::Array => None,
         }
         .ok_or_else(|| {
             ApiError::InternalServerError(format!(
                 "Computed sort field '{}' does not match its declared result type",
-                computed_sort.key()
+                computed_query.key()
             ))
         })
     }
